+++ resources/views/errors/403.fortune.php
@@ -1,8 +1,0 @@
-<% extends("errors/Error") %>
-
-<% part("errorTitle") %>
-    Forbidden
-<% endpart %>
-<% part("errorDescription") %>
-    You do not have permission to view this page.
-<% endpart %>