+++ resources/views/errors/404.fortune.php
@@ -1,8 +1,0 @@
-<% extends("errors/Error") %>
-
-<% part("errorTitle") %>
-    Page not found
-<% endpart %>
-<% part("errorDescription") %>
-    The link you clicked on appears to be broken.
-<% endpart %>