+++ resources/views/errors/html/401.fortune.php
@@ -1,0 +1,8 @@
+<% extends("errors/html/Error") %>
+
+<% part("errorTitle") %>
+Unauthorized
+<% endpart %>
+<% part("errorMessage") %>
+You do not have permission to view this page.
+<% endpart %>