+++ resources/views/errors/html/404.fortune.php
@@ -1,0 +1,8 @@
+<% extends("errors/html/Error") %>
+
+<% part("errorTitle") %>
+Page not found
+<% endpart %>
+<% part("errorMessage") %>
+The link you clicked on appears to be broken.
+<% endpart %>