+++ resources/views/errors/html/503.fortune.php
@@ -1,0 +1,8 @@
+<% extends("errors/html/Error") %>
+
+<% part("errorTitle") %>
+Down for maintenance
+<% endpart %>
+<% part("errorMessage") %>
+Please check back later.
+<% endpart %>