+++ resources/views/errors/json/400.fortune.php
@@ -1,0 +1,1 @@
+<% include("errors/json/Error", ["__errorCode" => 400, "__errorMessage" => "Your request was invalid"]) %>