+++ resources/views/errors/json/403.fortune.php
@@ -1,0 +1,1 @@
+<% include("errors/json/Error", ["__errorCode" => 403, "__errorMessage" => "You do not have permission for this resource"]) %>