+++ resources/views/errors/json/404.fortune.php
@@ -1,0 +1,1 @@
+<% include("errors/json/Error", ["__errorCode" => 404, "__errorMessage" => "The resource does not exist"]) %>