+++ resources/views/errors/json/503.fortune.php
@@ -1,0 +1,1 @@
+<% include("errors/json/Error", ["__errorCode" => 503, "__errorMessage" => "Please check back later"]) %>