+++ resources/views/errors/json/Error.fortune.php
@@ -1,0 +1,1 @@
+{{! json_encode(["error" => ["code" => $__errorCode, "message" => $__errorMessage]]) !}}