+++ src/Project/Bootstrappers/Http/Views/BuildersBootstrapper.php
@@ -4,7 +4,7 @@ namespace Project\Bootstrappers\Http\Views;
 use Opulence\Bootstrappers\Bootstrapper;
 use Opulence\Views\Factories\IViewFactory;
 use Opulence\Views\IView;
-use Project\Http\Views\Builders\ErrorBuilder;
+use Project\Http\Views\Builders\HtmlErrorBuilder;
 use Project\Http\Views\Builders\HomeBuilder;
 use Project\Http\Views\Builders\MasterBuilder;
 
@@ -26,8 +26,8 @@ class BuildersBootstrapper extends Bootstrapper
         $viewFactory->registerBuilder("Home", function (IView $view) {
             return (new HomeBuilder())->build($view);
         });
-        $viewFactory->registerBuilder("errors/Error", function (IView $view) {
-            return (new ErrorBuilder())->build($view);
+        $viewFactory->registerBuilder("errors/html/Error", function (IView $view) {
+            return (new HtmlErrorBuilder())->build($view);
         });
     }
 }